refactor the location of the Input component in the directory and rename the Login name folder
